+++ ReleaseNotes.md
@@ -5,6 +5,10 @@
 
 ##### Release log
 
+- 2.15.0: Changed the user interface for the entry page cards to have a tabbed look and responsive 1 or 2 columns
+- 2.15.0: Team scoreboards now allow overriding the SMHF Sinclair and Age Factor options
+- 2.15.0: Tracker will now use Sinclair 2028 if that is selected in OWLMCMS (but SMHF has its own specific options)
+
 - 2.14.2: team scoreboard toggle between fixed team order and ranking order now working
 - 2.14.1: team scoreboards result exports now include membership and body weight
 - 2.14.0: Remove hard-coded timer values, obey directives from owlcms